+++ Gemfile
@@ -28,6 +28,7 @@ gem 'devise-bootstrap-views','~> 1.0'
 gem 'iex-ruby-client'
 gem 'ransack'
 gem 'google_search_results'
+gem 'rubocop', '~> 1.19', require: false
 
 # Reduces boot times through caching; required in config/boot.rb
 gem 'bootsnap', '>= 1.4.4', require: false
