+++ app/assets/stylesheets/custom.css.scss
@@ -163,29 +163,28 @@ form.new_user{
 }
 
 #img-result{
-    margin-top: 1%;
-    float: left;
-    clear: right;
-    margin-right: 2%;
+	margin-top: 1%;
+	float: left;
+	clear: right;
+	margin-right: 2%;
 }
 
 #remaining-results{
-    margin-left: 12%;
+  margin-left: 12%;
 }
 
 .google-classic{
-    color: blue;
+  color: blue;
 }
 
 #source-result{
-    color: green;
+  color: green;
 }
 
 .input-margin{
-    margin-left: 1%;
+  margin-left: 1%;
 }
 
 i.fa { 
-    font-family:FontAwesome!important; 
+  font-family:FontAwesome!important; 
 }
-
