+++ app/controllers/stocks_controller.rb
@@ -1,25 +1,25 @@
 class StocksController < ApplicationController
-    
-    def search
-        if params[:stock].present?
-            client = Stock.conn()
-            @stock = Stock.new_lookup(client, params[:stock])
-            if @stock
-                respond_to do |format|
-                    format.js { render partial: 'users/sessions/result'}
-                end
-            else
-                respond_to do |format|
-                    flash.now[:alert] = "Please enter a valid symbol to search"
-                    format.js { render partial: 'users/sessions/result'}
-                end
-            end
-        else
-            respond_to do |format|
-                flash.now[:alert] = "Please enter a symbol to search"
-                format.js { render partial: 'users/sessions/result'}
-            end
-        end
-    end
 
+	def search
+		if params[:stock].present?
+			client = Stock.conn()
+			@stock = Stock.new_lookup(client, params[:stock])
+			if @stock
+				respond_to do |format|
+					format.js { render partial: 'users/sessions/result' }
+				end
+			else
+				respond_to do |format|
+					flash.now[:alert] = "Please enter a valid symbol to search"
+					format.js { render partial: 'users/sessions/result' }
+				end
+			end
+		else
+			respond_to do |format|
+				flash.now[:alert] = "Please enter a symbol to search"
+				format.js { render partial: 'users/sessions/result' }
+			end
+		end
+	end
+	
 end
