+++ app/controllers/users/passwords_controller.rb
@@ -1,11 +1,11 @@
 class Users::PasswordsController < Devise::PasswordsController
-    include ApplicationHelper
-    
-    def new
-        @new_pass = true
-        super
-    end 
+	include ApplicationHelper
+	
+	def new
+		@new_pass = true
+		super
+	end 
 
-    def update
-    end 
+	def update
+	end 
 end
