+++ app/controllers/users/sessions_controller.rb
@@ -1,6 +1,6 @@
-# frozen_string_literal: true
-
 class Users::SessionsController < Devise::SessionsController
+  before_action :set_search
+
   def new
     @new_user = true
     super
@@ -21,22 +21,24 @@ class Users::SessionsController < Devise::SessionsController
   end
 
   def search
-    if User.params_allowed?(params)
-      @friends = current_user.find_friends(params)
-        unless @friends == nil
-          respond_to do |format|
-          format.js { render partial: 'users/sessions/friend_result'}
-          end
-        else
-          respond_to do |format|
-            flash.now[:alert] = "No user found!"
-            format.js { render partial: 'users/sessions/friend_result'}
-          end
+    if params[:q].values.reject(&:blank?).any?
+      @q = User.ransack(params[:q])
+      friends =@q.result(distinct: true)
+      unless friends.empty?
+        @friends = friends.reject { |user| user.id == current_user.id }
+        respond_to do |format|
+          format.js { render partial: 'users/sessions/friend_result' }
         end
+      else
+        respond_to do |format|
+          flash.now[:alert] = "No user found!"
+          format.js { render partial: 'users/sessions/friend_result' }
+        end
+      end
     else
       respond_to do |format|
         flash.now[:alert] = "Please enter a friend's name or email to search"
-        format.js { render partial: 'users/sessions/friend_result'}
+        format.js { render partial: 'users/sessions/friend_result' }
       end
     end
   end
@@ -47,14 +49,13 @@ class Users::SessionsController < Devise::SessionsController
     if @tracked_stocks
       respond_to do |format|
         flash.now[:notice] = "Stocks updated succesfully!"
-        format.js {render partial: 'users/sessions/stocks/list'}
+        format.js {render partial: 'users/sessions/stocks/list' }
       end  
     else
       respond_to do |format|
         flash.now[:alert] = "Stocks can't be updated at the moment. Please try again!"
-        format.js {render partial: 'users/sessions/stocks/list'}
+        format.js {render partial: 'users/sessions/stocks/list' }
       end
-      
     end
   end
 
@@ -71,36 +72,18 @@ class Users::SessionsController < Devise::SessionsController
   end 
 
   def web_search
-    
     @search = GoogleSearch.new({
       q: params[:query], # search search
       tbm: "nws", # news
       tbs: "qdr:d", # last 24h
       num: 10
     })
-    
   end
-  # before_action :configure_sign_in_params, only: [:create]
 
-  # GET /resource/sign_in
-  # def new
-  #   super
-  # end
+  protected
+    def set_search
+      @q = User.ransack(params[:q])
+      GoogleSearch.api_key = "ff2bccdbce4b28080da429282d777be0bd266be8e41a57f3e3be001b59255653"
+    end
 
-  # POST /resource/sign_in
-  # def create
-  #   super
-  # end
-
-  # DELETE /resource/sign_out
-  # def destroy
-  #   super
-  # end
-
-  # protected
-
-  # If you have extra params to permit, append them to the sanitizer.
-  # def configure_sign_in_params
-  #   devise_parameter_sanitizer.permit(:sign_in, keys: [:attribute])
-  # end
 end
