+++ app/controllers/welcome_controller.rb
@@ -1,4 +1,6 @@
 class WelcomeController < ApplicationController
+  
   def index
   end
+
 end
