+++ app/helpers/application_helper.rb
@@ -1,13 +1,19 @@
 module ApplicationHelper
-    def resource_name
-        :user
-      end
+
+  def resource_name
+    :user
+  end
     
-      def resource
-        @resource ||= User.new
-      end
-    
-      def devise_mapping
-        @devise_mapping ||= Devise.mappings[:user]
-      end
+  def resource
+    @resource ||= User.new
+  end
+
+  def devise_mapping
+    @devise_mapping ||= Devise.mappings[:user]
+  end
+
+  def nav_status(path_a, path_b)
+    return 'active' if path_a == path_b
+  end
+
 end
