--- app/helpers/friendships_helper.rb
@@ -1,2 +1,0 @@
-module FriendshipsHelper
-end
