--- app/helpers/my_portfolio_helper.rb
@@ -1,5 +1,0 @@
-module MyPortfolioHelper
-    def nav_status(path_a, path_b)
-        return 'active' if path_a == path_b
-    end 
-end
