--- app/helpers/user_stocks_helper.rb
@@ -1,2 +1,0 @@
-module UserStocksHelper
-end
