--- app/helpers/welcome_helper.rb
@@ -1,2 +1,0 @@
-module WelcomeHelper
-end
