+++ app/views/layouts/_background.html.erb
@@ -1,7 +1,7 @@
 <% if @new_user %>
-        "new_user"
+  "new_user"
 <% elsif @sign_up %>   
-        "new_account"
+  "new_account"
 <% elsif @new_pass %>   
-        "new_pass"
+  "new_pass"
 <% end %>
