+++ app/views/layouts/_messages.html.erb
@@ -1,14 +1,16 @@
 <div class="container">
-<% if notice %>
-    <div class='alert alert-success'>
-        <button type="button" class="close" data-dismiss="alert" aria-label="Close"> <span aria-hidden="true">&times;</span> </button>
-        <%= notice %>
-    </div>
-    <% end %>
-<% if alert %>
-    <div class='alert alert-danger'>
-        <button type="button" class="close" data-dismiss="alert" aria-label="Close"> <span aria-hidden="true">&times;</span> </button>
-        <%= alert %>
-    </div>
-<% end %>
+	<% if notice %>
+		<div class='alert alert-success'>
+			<button type="button" class="close" data-dismiss="alert" aria-label="Close"> <span aria-hidden="true">&times;</span> </button>
+			<%= notice %>
+		</div>
+		
+	<% end %>
+	<% if alert %>
+		<div class='alert alert-danger'>
+			<button type="button" class="close" data-dismiss="alert" aria-label="Close"> <span aria-hidden="true">&times;</span> </button>
+			<%= alert %>
+		</div>
+
+	<% end %>
 </div>
