+++ app/views/users/registrations/edit.html.erb
@@ -28,7 +28,6 @@
             <%= f.label :password %> <i>(leave blank if you don't want to change it)</i><br />
             <%= f.password_field :password, class:'form-control', autocomplete: "new-password" %>
             <% if @minimum_password_length %>
-            
               <em><%= @minimum_password_length %> characters minimum</em>
             <% end %>
           </div>
@@ -52,7 +51,6 @@
             <h4>Cancel my account</h4>
             <p>Unhappy? <%= link_to "Cancel my account", registration_path(resource_name), class: 'btn btn-danger', data: { confirm: "Are you sure?" }, method: :delete %></p>
           </div>
-          
         <% end %>
       </div>
     </div>
