+++ app/views/users/registrations/new.html.erb
@@ -5,48 +5,47 @@
 
 <div id="signup" class="card text-center">
 	<div id = "signup-card-header" class="card-header"> Sign Up </div>
-	<div class="card-body">
-		<h5 class="card-title">Create an account</h5>
-		<%= form_for(resource, as: resource_name, url: registration_path(resource_name)) do |f| %>
-      <%= render "users/shared/error_messages", resource: resource %>
-      <div class="field">
-        <%= f.label :first_name %>
-        <br />
-        <%= f.text_field :first_name, autofocus: true, autocomplete: "first_name" %>
-      </div>
+    <div class="card-body">
+      <h5 class="card-title">Create an account</h5>
+      <%= form_for(resource, as: resource_name, url: registration_path(resource_name)) do |f| %>
+        <%= render "users/shared/error_messages", resource: resource %>
+        <div class="field">
+          <%= f.label :first_name %>
+          <br />
+          <%= f.text_field :first_name, autofocus: true, autocomplete: "first_name" %>
+        </div>
 
-      <div class="field">
-        <%= f.label :last_name %>
-        <br />
-        <%= f.text_field :last_name, autofocus: true, autocomplete: "last_name" %>
-      </div>
+        <div class="field">
+          <%= f.label :last_name %>
+          <br />
+          <%= f.text_field :last_name, autofocus: true, autocomplete: "last_name" %>
+        </div>
 
-      <div class="field">
-        <%= f.label :email %>
-        <br />
-        <%= f.email_field :email, autofocus: true, autocomplete: "email" %>
-      </div>
+        <div class="field">
+          <%= f.label :email %>
+          <br />
+          <%= f.email_field :email, autofocus: true, autocomplete: "email" %>
+        </div>
 
-      <div class="field">
-        <%= f.label :password %>
-        <% if @minimum_password_length %> <em>(<%= @minimum_password_length %> characters minimum)</em>
-        <% end %>
-        <br />
-        <%= f.password_field :password, autocomplete: "new-password" %>
-      </div>
+        <div class="field">
+          <%= f.label :password %>
+          <% if @minimum_password_length %> <em>(<%= @minimum_password_length %> characters minimum)</em>
+          <% end %>
+          <br />
+          <%= f.password_field :password, autocomplete: "new-password" %>
+        </div>
 
-      <div class="field">
-        <%= f.label :password_confirmation %>
-        <br />
-        <%= f.password_field :password_confirmation, autocomplete: "new-password" %>
-      </div>
+        <div class="field">
+          <%= f.label :password_confirmation %>
+          <br />
+          <%= f.password_field :password_confirmation, autocomplete: "new-password" %>
+        </div>
 
-      <div class="actions-signup">
-        <br>
-        <%= f.submit "Sign up", :class => "btn btn-success" %>
-        <%= render "users/shared/links" %>
-      </div>
-
-		<% end %>
+        <div class="actions-signup">
+          <br>
+          <%= f.submit "Sign up", :class => "btn btn-success" %>
+          <%= render "users/shared/links" %>
+        </div>
+      <% end %>
   </div>
 </div>
