+++ app/views/users/sessions/_friend_result.html.erb
@@ -1,26 +1,30 @@
 <div class="results-block">
 	<%= render 'layouts/messages' %>
 </div>
+
 <% if @friends %>
-    <div class="card card-header results-block">
-        <table class="col-md-12">
-            <tbody>
-                <% @friends.each do |friend| %>
-                    <tr>
-                        <td><strong>Name: </strong> <%= friend.full_name %> </td>
-                        <td><strong>Email: </strong> <%= friend.email %> </td>
-                        <td> <strong>Actions: </strong> </td>
-                        <td> <%= link_to 'Visit profile', view_profile_path(id: friend.id), class: "btn btn-success" %>
-                        <% if current_user.not_friends_with?(friend.id)%>
-                            <%= link_to 'Follow friend?', friendships_path(friend: friend), class: "btn btn-success", method: :post %>
-                        <% else %>
-                            <span class="badge badge-secondary"> You are friends </span>
-                        <% end %>
-                        </td>
-                    </tr>
-                <% end %>
-            </tbody>
-        </table>
-    </div>
+	<div class="card card-header results-block">
+		<table class="col-md-12">
+			<tbody>
+				<% @friends.each do |friend| %>
+					<tr>
+						<td><strong>Name: </strong> <%= friend.full_name %> </td>
+						<td><strong>Email: </strong> <%= friend.email %> </td>
+						<td> <strong>Actions: </strong> </td>
+
+						<td> 
+							<%= link_to 'Visit profile', view_profile_path(id: friend.id), class: "btn btn-success" %>
+							<% if current_user.not_friends_with?(friend.id)%>
+									<%= link_to 'Follow friend?', friendships_path(friend: friend), class: "btn btn-success", method: :post %>
+							<% else %>
+									<span class="badge badge-secondary"> You are friends </span>
+							<% end %>
+						</td>
+
+					</tr>
+				<% end %>
+			</tbody>
+		</table>
+	</div>
 
 <% end %>
