+++ app/views/users/sessions/_result.html.erb
@@ -3,23 +3,27 @@
 </div>
 
 <% if @stock %>
-    <div class="card card-header results-block">
-        <strong>Symbol: </strong> <%= @stock.ticker %>
-        <strong>Name: </strong> <%= @stock.name %>
-        <strong>Price: </strong> <%= @stock.last_price %>
-        <% if current_user.can_track_stock?(@stock.ticker) %>
-            <%= link_to 'Add to portfolio', user_stocks_path(user: current_user, ticker: @stock.ticker), class: "btn btn-success", method: :post %>
-        <% else%>
-            <span class="badge badge-secondary">
-                You are already tracking
-                <% if !current_user.under_stock_limit? %>
-                    10 stocks 
-                <% end %>
-                <% if current_user.stock_already_tracked?(@stock.ticker) %>
-                    this stock
-                <% end %>
-            </span>
-        <%end %>
-    </div>
-    
+	<div class="card card-header results-block">
+		<strong>Symbol: </strong> <%= @stock.ticker %>
+		<strong>Name: </strong> <%= @stock.name %>
+		<strong>Price: </strong> <%= @stock.last_price %>
+
+		<% if current_user.can_track_stock?(@stock.ticker) %>
+			<%= link_to 'Add to portfolio', user_stocks_path(user: current_user, ticker: @stock.ticker), class: "btn btn-success", method: :post %>
+		<% else%>
+			<span class="badge badge-secondary">
+			You are already tracking
+
+			<% if !current_user.under_stock_limit? %>
+				10 stocks 
+			<% end %>
+
+			<% if current_user.stock_already_tracked?(@stock.ticker) %>
+				this stock
+			<% end %>
+
+			</span>
+		<%end %>
+
+	</div>
 <% end %>
