+++ app/views/users/sessions/friends/_friend_list.html.erb
@@ -1,28 +1,32 @@
 <div class="row">
   <div class="col-9">
-      <table class="table table-bordered table-hover">
-        <thead>
-            <tr>
-              <th scope="col">Email</th>
-              <th scope="col">First Name</th>
-              <th scope="col">Last Name</th>
-              <th scope="col">Actions</th>
-            </tr>
-        </thead>
-        <tbody>
-            <% @friends.each do |friend| %>
-                <tr>
-                  <th scope="row"><%= friend.email %></th>
-                  <td><%= friend.first_name %></td>
-                  <td><%= friend.last_name %></td>
-                  <td> 
-                      <%= link_to 'View profile', view_profile_path(id: friend.id), class: 'btn btn-sm btn-success' %>
-                      <%= link_to 'Stop following', friendship_path(friend), method: :delete, data: {confirm: 'Are you sure?'}, class: 'btn btn-sm btn-warning' %>
-                  </td>
-                </tr>
-            <% end %>
-        </tbody>
-      </table>
+    <table class="table table-bordered table-hover">
+      <thead>
+        <tr>
+          <th scope="col">Email</th>
+          <th scope="col">First Name</th>
+          <th scope="col">Last Name</th>
+          <th scope="col">Actions</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        <% @friends.each do |friend| %>
+          <tr>
+            <th scope="row"><%= friend.email %></th>
+            <td><%= friend.first_name %></td>
+            <td><%= friend.last_name %></td>
+
+            <td> 
+              <%= link_to 'View profile', view_profile_path(id: friend.id), class: 'btn btn-sm btn-success' %>
+              <%= link_to 'Stop following', friendship_path(friend), method: :delete, data: { confirm: 'Are you sure?' }, class: 'btn btn-sm btn-warning' %>
+            </td>
+
+          </tr>
+        <% end %>
+      </tbody>
+
+    </table>
   </div>
 
 </div>
