+++ app/views/users/sessions/new.html.erb
@@ -1,38 +1,40 @@
 <div class="login-title">
-    <h1> Finance Tracker </h1>
-    <p> You app for finance </p>
+	<h1> Finance Tracker </h1>
+	<p> You app for finance </p>
 </div>
 
 <div id="login" class="card text-center">
-    <div class="card-header"> Login </div>
-    <div class="card-body">
-        <h5 class="card-title">Enter your credentials</h5>
-        <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
-            <div class="field">
-                <%= f.label :email %>
-                <br />
-                <%= f.email_field :email, autofocus: true, autocomplete: "email" %>
-            </div>
+	<div class="card-header"> Login </div>
+	<div class="card-body">
+		<h5 class="card-title">Enter your credentials</h5>
+		<%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
 
-            <div class="field">
-                <%= f.label :password %>
-                <br />
-                <%= f.password_field :password, autocomplete: "current-password" %>
-            </div>
+			<div class="field">
+				<%= f.label :email %>
+				<br />
+				<%= f.email_field :email, autofocus: true, autocomplete: "email" %>
+			</div>
 
-            <% if devise_mapping.rememberable? %>
-                <div class="field">
-                    <%= f.check_box :remember_me %>
-                    <%= f.label :remember_me %>
-                </div>
+			<div class="field">
+				<%= f.label :password %>
+				<br />
+				<%= f.password_field :password, autocomplete: "current-password" %>
+			</div>
 
-            <% end %>
-            <div class="actions-login">
-                <%= f.submit "Log in",:class => "btn btn-primary" %>
-            </div>
+			<% if devise_mapping.rememberable? %>
 
-        <% end %>
-        <%= render "users/shared/links" %>
-    </div>
-    
+				<div class="field">
+					<%= f.check_box :remember_me %>
+					<%= f.label :remember_me %>
+				</div>
+
+			<% end %>
+
+			<div class="actions-login">
+				<%= f.submit "Log in",:class => "btn btn-primary" %>
+			</div>
+
+		<% end %>
+		<%= render "users/shared/links" %>
+	</div>
 </div>
