+++ app/views/users/sessions/show.html.erb
@@ -1,27 +1,29 @@
 <div class="container">
-    <h2 class="main-title"> <%= "#{@user.full_name}'s profile" %> </h2>
+	<h2 class="main-title"> <%= "#{@user.full_name}'s profile" %> </h2>
 
-    <div class="card card-header results-block">
-        <table class="col-md-12">
-            <tbody>
-                <tr>
-                    <td><strong>Name: </strong> <%= @user.full_name %> </td>
-                    <td><strong>Email: </strong> <%= @user.email %> </td>
-                    <td> 
-                        <% if current_user.not_friends_with?(@user.id)%>
-                            <%= link_to 'Follow friend?', friendships_path(friend: @user), class: "btn btn-success", method: :post %>
-                        <% else %>
-                            <span class="badge badge-secondary"> You are already friends </span>
-                        <% end %>
-                    </td>
-                </tr>
-            </tbody>
-        </table>
-    </div>
+	<div class="card card-header results-block">
+		<table class="col-md-12">
+			<tbody>
+				<tr>
+					<td><strong>Name: </strong> <%= @user.full_name %> </td>
+					<td><strong>Email: </strong> <%= @user.email %> </td>
 
-    <% unless @tracked_stocks.empty? %>
-        <%= render 'users/sessions/stocks/list' %>
-    <% else %>
-        <p class='lead'> This user is not tracking any stock </p>
-    <% end %>
+					<td> 
+							<% if current_user.not_friends_with?(@user.id)%>
+									<%= link_to 'Follow friend?', friendships_path(friend: @user), class: "btn btn-success", method: :post %>
+							<% else %>
+									<span class="badge badge-secondary"> You are already friends </span>
+							<% end %>
+					</td>
+					
+				</tr>
+			</tbody>
+		</table>
+	</div>
+
+	<% unless @tracked_stocks.empty? %>
+			<%= render 'users/sessions/stocks/list' %>
+	<% else %>
+			<p class='lead'> This user is not tracking any stock </p>
+	<% end %>
 </div>
