+++ app/views/users/sessions/stocks/_list.html.erb
@@ -1,51 +1,59 @@
 <div class="results-block">
 	<%= render 'layouts/messages' %>
 </div>
+
 <div class="row">
   <div class="col-9">
     <table class="table table-bordered table-hover">
       <thead>
-          <tr>
-            <th scope="col">Ticker</th>
-            <th scope="col">Name</th>
-            <th scope="col">Price 
-              <span id ="price-buttons"> 
-                <%= link_to '', refresh_path, remote: true,  class: "btn btn-success fa fa-redo-alt" %> 
-                <%= link_to '', order_up_path, remote: true, class: "btn btn-success fa fa-angle-double-up" %>
-                <%= link_to '', order_down_path, remote: true, class: "btn btn-success fa fa-angle-double-down" %>
-              </span>
-            </th>
-            <th scope="col">Actions</th>
-          </tr>
+        <tr>
+          <th scope="col">Ticker</th>
+          <th scope="col">Name</th>
+
+          <th scope="col">
+            Price 
+            <span id ="price-buttons"> 
+              <%= link_to '', refresh_path, remote: true,  class: "btn btn-success fa fa-redo-alt" %> 
+              <%= link_to '', order_up_path, remote: true, class: "btn btn-success fa fa-angle-double-up" %>
+              <%= link_to '', order_down_path, remote: true, class: "btn btn-success fa fa-angle-double-down" %>
+            </span>
+          </th>
+
+          <th scope="col">Actions</th>
+        </tr>
       </thead>
 
       <tbody>
-          <% @tracked_stocks.each do |stock| %>
-              <tr>
-                <th scope="row"><%= stock.ticker %></th>
-                <td><%= stock.name %></td>
-                <td><%= stock.last_price %> </td>
-                <td>
-                  <% if current_user.can_track_stock?(stock.ticker) %>
-                    <%= link_to 'Add to portfolio', user_stocks_path(user: current_user, ticker: stock.ticker), class: "btn btn-success", method: :post %>
-                  <% elsif current_user.id != @user.id %>
-                    <span class="badge badge-secondary">
-                      You are already tracking
-                      <% if !current_user.under_stock_limit? %>
-                          10 stocks 
-                      <%# end %>
-                      <% elsif current_user.stock_already_tracked?(stock.ticker) %>
-                          this stock
-                      <% end %>
-                    </span>
-                  <%end %> 
-                  <% if current_user.id == @user.id %>
-                    <%= link_to 'Remove', user_stock_path(stock), method: :delete, data: {confirm: 'Are you sure?'}, class: 'btn btn-sm btn-warning' %>
+        <% @tracked_stocks.each do |stock| %>
+          <tr>
+            <th scope="row"><%= stock.ticker %></th>
+            <td><%= stock.name %></td>
+            <td><%= stock.last_price %></td>
+
+            <td>
+              <% if current_user.can_track_stock?(stock.ticker) %>
+                <%= link_to 'Add to portfolio', user_stocks_path(user: current_user, ticker: stock.ticker), class: "btn btn-success", method: :post %>
+              <% elsif current_user.id != @user.id %>
+                <span class="badge badge-secondary">
+                  You are already tracking
+                  <% if !current_user.under_stock_limit? %>
+                      10 stocks 
+                  <%# end %>
+                  <% elsif current_user.stock_already_tracked?(stock.ticker) %>
+                      this stock
                   <% end %>
-                </td>
-              </tr>
-          <% end %>
+                </span>
+              <%end %>
+
+              <% if current_user.id == @user.id %>
+                <%= link_to 'Remove', user_stock_path(stock), method: :delete, data: { confirm: 'Are you sure?' }, class: 'btn btn-sm btn-warning' %>
+              <% end %>
+            </td>
+
+          </tr>
+        <% end %>
       </tbody>
+      
     </table>
   </div>
