+++ app/views/welcome/index.html.erb
@@ -1,7 +1,6 @@
 <div class="container">
     <h1 class="main-title">Hello, <%= current_user.full_name %></h1>
-    <p>Welcome to Finance Tracker app</p>
-    
+    <p>Welcome to Finance Tracker app</p>  
     <br>
     <h2> Latest </h2>
     <p> Google is leading right now is stocks </p>
